Completo lo que falta para lograr entrenar un MLP Agrego tensorboard logging Escalo las bboxes para que sean relativas a la imagen original

+++ src/catdog/catdog/dataset.py
@@ -11,7 +11,7 @@ class CatDogDataset(Dataset):
         self.width = cat_dog_df["width"].values
         self.height = cat_dog_df["height"].values
         self.target = np.where(cat_dog_df["class"].values == "cat", 1, 0)
-        self.bbox = cat_dog_df[["xmin", "ymin", "xmax", "ymax"]].values
+        self.bbox = cat_dog_df[["xmin", "ymin", "xmax", "ymax"]].values.astype(np.float32)
         self.resizer = fn.Resize(img_output_size)
 
         self.transforms = transforms
